Use ddof in `numbagg>=0.7.0` for aggregations (#8624)
* Use numbagg ddof for aggregations

(not yet for rolling)

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------
